backend/requirements.txt
sentence-transformers==2.7.0
langchain>=0.2.0
langgraph>=0.0.69
faiss-cpu==1.8.0
requests==2.31.0
beautifulsoup4==4.12.2
fastapi==0.104.1
uvicorn==0.24.0
numpy==1.26.0
scikit-learn==1.3.2
ollama>=0.6.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
pydantic-sqlalchemy==0.1.0